fix(5-1): move currentOffer from cities to offers
