+++ src/components/app/app.jsx
@@ -4,8 +4,8 @@ import {BrowserRouter as Router, Route, Switch} from "react-router-dom";
 import {connect} from 'react-redux';
 
 import {ActionCreator} from "../../store/cities/actions.js";
-import {getCurrentOffers} from "../../store/offers/selectors.js";
-import {getCurrentOffer, getCurrentCity, getCities} from "../../store/cities/selectors.js";
+import {getCurrentOffers, getCurrentOffer} from "../../store/offers/selectors.js";
+import {getCurrentCity, getCities} from "../../store/cities/selectors.js";
 
 import {CITIES} from "../../utils/constants.js";
 import {offerShape} from "../../utils/prop-types.js";
