+++ src/store/cities/cities.test.js
@@ -1,7 +1,6 @@
 import {ActionCreator, ActionType} from "./actions.js";
 import {reducer} from "./reducer.js";
 
-import {testOffer} from "../../mocks/test-mocks.js";
 import {CITIES} from "../../utils/constants.js";
 
 describe(`Reducer work correctly`, () => {
@@ -9,7 +8,6 @@ describe(`Reducer work correctly`, () => {
     expect(reducer(undefined, {})).toEqual({
       cities: CITIES,
       currentCity: CITIES[0],
-      currentOffer: null,
     });
   });
 
@@ -17,29 +15,12 @@ describe(`Reducer work correctly`, () => {
     expect(reducer({
       cities: CITIES,
       currentCity: CITIES[0],
-      currentOffer: null,
     }, {
       type: ActionType.SET_CURRENT_CITY,
       payload: CITIES[1],
     })).toEqual({
       cities: CITIES,
       currentCity: CITIES[1],
-      currentOffer: null,
-    });
-  });
-
-  it(`Reducer should set current offer by a given value`, () => {
-    expect(reducer({
-      cities: CITIES,
-      currentCity: CITIES[0],
-      currentOffer: null,
-    }, {
-      type: ActionType.SET_CURRENT_OFFER,
-      payload: testOffer,
-    })).toEqual({
-      cities: CITIES,
-      currentCity: CITIES[0],
-      currentOffer: testOffer,
     });
   });
 });
@@ -49,13 +30,6 @@ describe(`Action creators work correctly`, () => {
     expect(ActionCreator.setCurrentCity(CITIES[1])).toEqual({
       type: ActionType.SET_CURRENT_CITY,
       payload: CITIES[1],
-    });
-  });
-
-  it(`Action creator for loadOffers returns correct action`, () => {
-    expect(ActionCreator.setCurrentOffer(testOffer)).toEqual({
-      type: ActionType.SET_CURRENT_OFFER,
-      payload: testOffer,
     });
   });
 });
