+++ src/store/cities/reducer.js
@@ -5,7 +5,6 @@ import {ActionType} from "./actions.js";
 const initialState = {
   cities: CITIES,
   currentCity: CITIES[0],
-  currentOffer: null,
 };
 
 export const reducer = (state = initialState, action) => {
@@ -13,11 +12,6 @@ export const reducer = (state = initialState, action) => {
     case ActionType.SET_CURRENT_CITY:
       return extendObject(state, {
         currentCity: action.payload,
-      });
-
-    case ActionType.SET_CURRENT_OFFER:
-      return extendObject(state, {
-        currentOffer: action.payload,
       });
 
     default: return state;
