+++ src/store/offers/offers.test.js
@@ -1,22 +1,38 @@
 import {ActionCreator, ActionType} from "./actions.js";
 import {reducer} from "./reducer.js";
-import {testOffers} from "../../mocks/test-mocks.js";
+import {testOffers, testOffer} from "../../mocks/test-mocks.js";
 
 describe(`Reducer work correctly`, () => {
   it(`Reducer without additional parameters should return initial state`, () => {
     expect(reducer(undefined, {})).toEqual({
       offers: [],
+      currentOffer: null,
     });
   });
 
   it(`Reducer should load offers by a given value`, () => {
     expect(reducer({
       offers: [],
+      currentOffer: null,
     }, {
       type: ActionType.LOAD_OFFERS,
       payload: testOffers,
     })).toEqual({
       offers: testOffers,
+      currentOffer: null,
+    });
+  });
+
+  it(`Reducer should set current offer by a given value`, () => {
+    expect(reducer({
+      offers: [],
+      currentOffer: null,
+    }, {
+      type: ActionType.SET_CURRENT_OFFER,
+      payload: testOffer,
+    })).toEqual({
+      offers: [],
+      currentOffer: testOffer,
     });
   });
 });
@@ -26,6 +42,13 @@ describe(`Action creators work correctly`, () => {
     expect(ActionCreator.loadOffers(testOffers)).toEqual({
       type: ActionType.LOAD_OFFERS,
       payload: testOffers,
+    });
+  });
+
+  it(`Action creator for loadOffers returns correct action`, () => {
+    expect(ActionCreator.setCurrentOffer(testOffer)).toEqual({
+      type: ActionType.SET_CURRENT_OFFER,
+      payload: testOffer,
     });
   });
 });
