+++ src/store/offers/reducer.js
@@ -3,6 +3,7 @@ import {ActionType} from "./actions.js";
 
 const initialState = {
   offers: [],
+  currentOffer: null,
 };
 
 export const reducer = (state = initialState, action) => {
@@ -10,6 +11,11 @@ export const reducer = (state = initialState, action) => {
     case ActionType.LOAD_OFFERS:
       return extendObject(state, {
         offers: action.payload,
+      });
+
+    case ActionType.SET_CURRENT_OFFER:
+      return extendObject(state, {
+        currentOffer: action.payload,
       });
 
     default: return state;
